Backend/app/requirements.txt
fastapi
uvicorn
python-jose
supabase
pydantic